`[Configure/Queues]` Add support for Additional Security Groups in Queues.
