+++ frontend/src/pages/Configure/HeadNode.js
@@ -17,7 +17,6 @@ import { findFirst, getIn } from '../../util'
 // UI Elements
 import {
   Box,
-  Button,
   ColumnLayout,
   ExpandableSection,
   FormField,
@@ -25,14 +24,13 @@ import {
   Select,
   SpaceBetween,
   Toggle,
-  TokenGroup,
 } from "@awsui/components-react";
 
 // State
 import { setState, getState, useState, clearState, updateState, ssmPolicy } from '../../store'
 
 // Components
-import { SubnetSelect, InstanceSelect, ActionsEditor } from './Components'
+import { ActionsEditor, InstanceSelect, SecurityGroups, SubnetSelect } from './Components'
 import HelpTooltip from '../../components/HelpTooltip'
 
 // Constants
@@ -199,37 +197,6 @@ function SsmSettings() {
   )
 }
 
-function SecurityGroups() {
-  const sgPath = [...headNodePath, 'Networking', 'AdditionalSecurityGroups'];
-  const selectedSgs = useState(sgPath) || [];
-  const sgSelected = useState(['app', 'wizard', 'sg-selected']);
-
-  const sgs = useState(['aws', 'security_groups']) || [];
-  // const sgMap = sgs.reduce((acc, s) => {acc[s.GroupId] = s.GroupName; return acc}, {})
-  console.log(useState(headNodePath));
-
-  const itemToOption = item => {return {value: item.GroupId, label: item.GroupId, description: item.GroupName}}
-  const removeSg = (i) => {
-    setState(sgPath, [...selectedSgs.slice(0, i), ...selectedSgs.slice(i + 1)]);
-    if(getState(sgPath).length === 0)
-      clearState(sgPath);
-  }
-  return <SpaceBetween direction="vertical" size="xs">
-    <div style={{display: "flex", flexDirection: "row", alignItems: "center", justifyContent: "space-between", gap:"16px"}}>
-      <Select
-        selectedOption={(sgSelected && findFirst(sgs, x => x.GroupId === sgSelected.value)) ? itemToOption(findFirst(sgs, x => x.GroupId === sgSelected.value)) : {label: "Please Select A Security Group"}}
-        onChange={({detail}) => {setState(['app', 'wizard', 'sg-selected'], detail.selectedOption)}}
-        options={sgs.map(itemToOption)}
-      />
-      <Button disabled={!sgSelected} onClick={() => setState(sgPath, [...selectedSgs, sgSelected.value])}>Add</Button>
-    </div>
-    <TokenGroup
-      onDismiss={({ detail: { itemIndex } }) => {removeSg(itemIndex)}}
-      items={selectedSgs.map((s) => {return {label: s, dismissLabel: `Remove ${s}`}})}
-    />
-  </SpaceBetween>
-}
-
 function DcvSettings() {
   const dcvPath = [...headNodePath, 'Dcv', 'Enabled'];
 
@@ -393,7 +360,7 @@ function HeadNode() {
         </div>
         <div key="sgs" style={{display: "flex", flexDirection: "row", alignItems: "center", justifyContent: "space-between"}}>
         <FormField label="Additional Security Groups">
-          <SecurityGroups />
+          <SecurityGroups basePath={headNodePath} />
         </FormField>
           <HelpTooltip>
             Provides additional security groups for the HeadNode.
